+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Features
 
 * Added the ability to [rotate authentication tokens](./Rotator.md)
+* Added the ability to bridge from YouTube Music
 
 ## Internal Changes